이미지 업로드 기능 개선/ DTO 개선(2)

+++ src/main/java/com/project/localbatter/components/DeleteFile.java
@@ -16,9 +16,7 @@ public class DeleteFile {
         for(String file : files){
             if(!file.contains("default")){
                 File f = new File(root + file);
-
-                if(f.exists())
-                    f.delete();
+                if(f.exists()) f.delete();
             }
         }
     }
